Process event updates
When identical events are emitted for a resource, kube-apiserver updates the previous
event and increments the count field to represent how many times an event occurred. This
is to optimize storage on ETCD. Such consolidation results in an `update` on the event.

Currently, the events exporter doesn't process any events that get updated so when similar
events re-occur, they aren't processed. This change enables processing events that get updated.